process_icell8.py: split out the pipeline infrastructure to 'pipeliner.py' module.
